refactor: convert device-screen class into helper function
